+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - General Release.
 
 ### Changed
+- load_cfg:  Added error messages to output return.
 - Removed unnecessary \*\*kwargs from argument list.
 - config/merge.py.TEMPLATE:  Set "merge-repo" as the base directory structure.
 - Documentation updates.